+++ src/ssl_stubs.c
@@ -1074,20 +1074,29 @@ CAMLprim value ocaml_ssl_version(value socket) {
 CAMLprim value ocaml_ssl_get_current_cipher(value socket) {
   CAMLparam1(socket);
   SSL *ssl = SSL_val(socket);
-
   caml_release_runtime_system();
   SSL_CIPHER *cipher = (SSL_CIPHER *)SSL_get_current_cipher(ssl);
   caml_acquire_runtime_system();
   if (!cipher)
     caml_raise_constant(*caml_named_value("ssl_exn_cipher_error"));
-
+#if defined(NO_NAKED_POINTERS) || defined(NAKED_POINTERS_CHECKER)
+  value vcipher = caml_alloc_shr(1, Abstract_tag);
+  *((SSL_CIPHER **) Data_abstract_val(vcipher)) = cipher;
+  CAMLreturn(vcipher);
+#else
   CAMLreturn((value)cipher);
+#endif
 }
 
 CAMLprim value ocaml_ssl_get_cipher_description(value vcipher) {
   CAMLparam1(vcipher);
   char buf[1024];
+
+#if defined(NO_NAKED_POINTERS) || defined(NAKED_POINTERS_CHECKER)
+  SSL_CIPHER *cipher = *((SSL_CIPHER **) Data_abstract_val(vcipher));
+#else
   SSL_CIPHER *cipher = (SSL_CIPHER *)vcipher;
+#endif
 
   caml_release_runtime_system();
   SSL_CIPHER_description(cipher, buf, 1024);
@@ -1099,7 +1108,12 @@ CAMLprim value ocaml_ssl_get_cipher_description(value vcipher) {
 CAMLprim value ocaml_ssl_get_cipher_name(value vcipher) {
   CAMLparam1(vcipher);
   const char *name;
+
+#if defined(NO_NAKED_POINTERS) || defined(NAKED_POINTERS_CHECKER)
+  SSL_CIPHER *cipher = *((SSL_CIPHER **) Data_abstract_val(vcipher));
+#else
   SSL_CIPHER *cipher = (SSL_CIPHER *)vcipher;
+#endif
 
   caml_release_runtime_system();
   name = SSL_CIPHER_get_name(cipher);
@@ -1111,7 +1125,12 @@ CAMLprim value ocaml_ssl_get_cipher_name(value vcipher) {
 CAMLprim value ocaml_ssl_get_cipher_version(value vcipher) {
   CAMLparam1(vcipher);
   const char *version;
+
+#if defined(NO_NAKED_POINTERS) || defined(NAKED_POINTERS_CHECKER)
+  SSL_CIPHER *cipher = *((SSL_CIPHER **) Data_abstract_val(vcipher));
+#else
   SSL_CIPHER *cipher = (SSL_CIPHER *)vcipher;
+#endif
 
   caml_release_runtime_system();
   version = SSL_CIPHER_get_version(cipher);